backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.8.2
    # via flask
click==8.1.7
    # via flask
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==4.0.1
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
werkzeug==3.0.3
    # via flask
openai
python-dotenv==1.0.1

